Create, Read, delete done,

+++ src/main/java/web/models/User.java
@@ -53,7 +53,7 @@ public class User {
         this.surname = lastName;
     }
 
-    public long getAge() {
+    public int getAge() {
         return age;
     }
